StateSelector and Filters are now .jsx filetype, troubleshot and fixed Filter/StateSelector to render correct list of usStates w/o error.
