+++ src/utils/states.js
@@ -51,4 +51,6 @@ const usStates = [
     { name: 'WYOMING', abbreviation: 'WY' }
 ]
 
-module.exports = usStates
+
+
+export default usStates
